requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
datetime-truncate==1.1.1
defusedxml==0.8.0rc2
Django==3.2.22
django-admin-charts==1.3.0
django-bower==5.2.0
django-cleanup==8.1.0
django-jet-reboot==1.3.3
django-light==0.1.0.post3
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-multiupload==0.6.1
django-nvd3==0.9.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mock==5.1.0
# mysqlclient==2.2.1
numpy==1.23.5
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
Unidecode==1.3.7
urllib3==2.1.0
Werkzeug==2.2.2
whitenoise==6.6.0
